rag/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my_rag_project"
version = "0.1.0"
description = "My RAG Learning Project"
dependencies = [
    "httptools>=0.7.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.9",
    "python-dotenv",
    "websockets>=16.0",
]

[tool.setuptools.packages.find]
where = ["."]